feat: Adds input validation
