+++ components/CognitiveMapping.tsx
@@ -1,0 +1,51 @@
+const functionToIndicesMap: Record<string, string[]> = {
+  "Fonctionnement intellectuel global": [
+    "QI",
+    "GAI",
+    "Compréhension verbale",
+    "Raisonnement perceptif",
+    "Mémoire de travail",
+    "Vitesse de traitement",
+  ],
+  "Fonctions exécutives": [
+    "Figure de Rey Rappel Immédiat",
+    "Figure de Rey Rappel Différé",
+    "Figure de Rey Temps de Copie",
+    "Figure de Rey Reconnaissance",
+    "Stroop D-Kefs Dénomination de couleur",
+    "Stroop D-Kefs Lecture de mots",
+    "Stroop D-Kefs Inhibition",
+    "Stroop D-Kefs Inhibition/Alternance",
+    "Tour D-Kefs Score de réussite total",
+    "Tour D-Kefs Temps moyen 1er mouvement",
+    "Tour D-Kefs Ratio temps-par-mouvement",
+    "Tour D-Kefs Ratio précision de mouvement",
+    "Tour D-Kefs Ratio violation de règle par item",
+    "Trail-Making Temps trail A",
+    "Trail-Making Temps trail B",
+    "Wisconsin Nombre d’essais",
+    "Wisconsin Nombre de réponse correcte",
+    "Wisconsin Nombre d’erreurs",
+    "Wisconsin Réponses persévératives",
+    "Wisconsin Erreurs persévératives",
+    "Wisconsin Erreurs non-persévératives",
+    "Wisconsin Niveau de réponses conceptuelles",
+    "Wisconsin Nombre de catégories complétées",
+    "Wisconsin Nombre d’essais pour compléter la première catégorie",
+    "Wisconsin Échec du maintien de catégorie",
+    "Wisconsin Apprendre à apprendre",
+  ],
+
+  Attention: [
+    "CPT-3 Détectabilité",
+    "CPT-3 Omissions",
+    "CPT-3 Commissions",
+    "CPT-3 Persévérations",
+    "CPT-3 HRT",
+    "CPT-3 Variabilité",
+    "CPT-3 HRT Block change",
+    "CPT-3 HRT ISI",
+  ],
+};
+
+export default functionToIndicesMap;
